requirements.txt
APScheduler==3.6.3
Flask==1.1.1
Flask-RESTful==0.3.8
lxml==4.5.0
numpy==1.19.2
pre-commit==2.7.1
psutil==5.7.2
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.12.0
requests==2.24.0
selenium==3.141.0
tqdm==4.50.2
wget==3.2

torch==1.5.0
torchvision==0.6.0
langdetect==1.0.8
Pillow==7.2.0
matplotlib==3.3.2

pytest==6.1.2
coverage==5.3